Rename GUI and Headless Browsers guide (breaks existing URL but we really need to do it)
Previous title was misleading because Selenium Server is not (typically) used on travis-ci.org.
Also, it was limiting the guide to just headless browsers while we have Clojure and Java GUI applications
tested on travis-ci.org and there is plenty of room for many kinds of test suites that may need
X Server/GUI, for example, visualization tools.