update Python and PHP version
Added PHP 5.5 and Python 3.3

I didn't add PHP 5.2. Maybe that was left off for a reason.

I changed the Python example from 3.2 to 3.3.
Many users might want to test the latest Python 2 and Python 3 only
and maybe Python 2.6 for the lowest version that makes it somewhat
easier to cater for Python 2 and 3.