Android: recommend to specify all dependencies
This practice ensure stable build environment even if some preinstalled
components are removed from the Travis VM. Unfortunately, it will also
increase the total build time (until we find a solution to this
problem).

See discussion in travis-ci/travis-ci#2296 for full details about the
advantages and drawbacks to rely on the pre-installed components.