Add commentary on 'pecl install'
I couldn't figure out how to recover from the error when
the desired version is already present on the system, so
no workaround is offered.

The exit status of 'pecl' is 1, and there is nothing noteworthy
in the commands to detect this condition and advise users.

The worst case scenario is that, when we update VM and installed
the most up-to-date package, 'pecl install' will start failing
randomly, and there is nothing we can do about it.

One may consider `pecl install X || true`, but then legitimate errors
from `pecl` will be ignored.