give more attention to Oracle JDK 8
Oracle JDK 8 is a production release and this PR gives more attention to
this being a mainstream release. I also changed/dropped the minor
versions as these will probably always be out-of-date.

Also I think it's safe to say "Support for multiple JDKs will be
available in the future." is an outdated statement.