Reference travis-build directly
The sentence referred to the first element in the list following it, however
this was unclear and could also be read "only one repository is required",
which leads the reader to wondering which repository is meant.  What was
meant was "only number 1 in the following list is required".  To make this
sense clearer, the name of the required repository is now used instead of a
number.