Make clearer how to update the travis-web project for matrix expansion
The pull request cited in the section about how to update the `travis-web`
project for language matrix expansion points to a commit to a file which no
longer exists.  It is thus hard to find the correct file to update.  This
change updates the text to mention the file which currently needs to be
edited in order to add a new language to the list of language configuration
keys, and thus hopefully reduces confusion on the part of the user.