No skin is removed.
`-no-skin` parameter removes the skin. This means that the emulator will be 320x480 pixel widthxheight and 320 density which messes up with most of the tests.
The default skin is 480x800 which is fine for most of the cases.

`--skin WXGA720` can be used like below to make tests more trusted. This is the best emulator config that is found by default.

    - echo no | android create avd --force -n test -t android-22 --abi armeabi-v7a --skin WXGA720
Conflicts:
	user/languages/android.md