Note MPI and $CC & $CXX interaction
https://github.com/travis-ci/travis-ci/issues/4682 reports
that MPI builds will have a hard time if
CC and CXX are set.

Changing default behavior is not practical, and sneaking around
the user intentions to deal with this special case is wrought
with edge cases.

So we will document, so that users may find it and shape
their builds to their liking.