Replace 'rbx-2' with 'rbx'.
Rubinius no longer supports any version of the 2.x epoch. Use of 'rbx' is
encouraged so that the current stable release is tested.

Rubinius releases as frequently as once or more a day and any attempt to
'lock' to a Rubinius version significantly disadvantages a project by not
using Rubinius fixes and enhancements.

It also disadvantages Rubinius because time is wasted dealing with outdated
issues for which fixes have already been released.