adding example
This example shows how to add the Travis CI Badge to a Readme in a Github Repository. I understand this may seem intuitive, but I think it's necessary to be explicit, so that new users can get up and running quickly. I found this document hard to digest as a new user to Travis.