caching: Travis CI fetches the cache of the default branch
The documentation states that "If a branch does not have its own
cache, Travis CI fetches the master branch cache".  However, this is
not quite true: it fetches the cache of the repository's default
branch, whatever it is called.  While the default branch is indeed
called 'master' in most repositories, that's not always the case.