Explain that services: docker only works on hosted (#1780)
* Explain that services: docker only works on hosted

We're seeing more and more customers running into this where they're trying to use Docker in their builds running on Enterprise. Though for Enterprise the setup is a bit different here.

* Improve wording in "how to Docker with Enterprise"

* It's sudo-enabled instead of sudo: required

* Better wording for "Here's how you do it on Enterprise"

* Have the enterprise note rendered separately

So it stands out and people actually read it. Otherwise it would follow
right after "We don't support OSX" without a line break

* Better display text for enterprise url

* Fix missing backticks for config file path

* Switch this around