Explain what to do if curl errors out during the build (#1888)
* Explain what to do if curl errors out during the build

We're seeing this kind of behavior together with self-signed
certificates, where the worker tries to fetch data from the platform
machine. Since the self-signed certificate is not trusted by curl, these
calls fail. Currently the only clean solution for this is to install a
certificate which got issued by a trusted CA.

* Minor text updates

* link fix