Rename Enterprise Operations Manual To Troubleshooting guide
Renames the page labeled 'Operations Manual' to 'Troubleshooting
Guide' as I feel it better fits what it is doing. It is also a
bit confusing to have the overall section be named 'Operations
Manual' and then have a specific page also named the same thing.

In addition this moves the 'backups' section into 'Platform
Administration Tips' since it doesn't fit with the rest of
the troubleshooting guide, which is mainly about solving bugs
or issues.