Because you should never do mvn install on any CI
Maven `install` phase is *not* about gathering the project dependencies and plugins binaries to store them locally. 
It's about generating the binary of the project and storing it in the `localRepository`, sort of local proxy for Maven artifacts.

The current documented command is in fact recommending to build the project in the TravisCI installation step.
Then build it again in the normal build step.

CI systems should *never ever* perform `mvn install` or any equivalent for any build and project configuration tool.