Use dir instead of local_dir
I tried today the travis v2 deploy to netfliy option and red this
documentation for it: https://docs.travis-ci.com/user/deployment-v2/providers/netlify/#deploying-a-specific-directory
In the part `known options` it mentions, that when I want to deploy a
specific folder it needs to be defined with the option `dir`
On the bottom of the documentation is an area `Deploying a specific
directory` that wants `local_dir` as a parameter. When I specify
`local_dir` it tells me that the option does not exist. With `dir` it
works.