Add rake task to update Ciphers list
Strictly speaking, it is necessary to consult the list of Ciphers
from the version of Ruby that is running `travis-tasks`, but
determining this is a little tricky (there is no API for such
information, say). So we will assume that the version of Ruby running
the docs site is reasonably close to that of `travis-tasks`, and
present the list this way.