Update footer.html to fix the status indicator dot
An additional div.status class was created in the css files for python and php pages. As both the pages end with a table of archived versions in which the status seemed to a be a declared object. I have made edits to the footer such that the indicator dot remains the same regardless of the changes made on the webpage.